Enable seeding the DB
ts-node is needed for compiling seed script on the flt.
Sadly, I had to get rid of path aliases, as ts-node couldn't
get the tsconfig working (a if it did not use the tsconfig at all?)

Idea is that I might get away with doubling the configuration both
in tsconfig and in package.json where ts-config is configured.