Use v4 of nanoid
The reason is that I can't use nanoid in v3 because Jest can't
figure out there's a CJS version of it and it loads the ESM one
and it can't use it as Jest can't speak ESM. So I had to mock
that module. So I can go with v4 that is shipped in ESM only,
as it doesn't matter to me anymore.