Remove all custom login mechanics and use 3rd party ones
GH for now. Twitter is not working ... docs, Google, etc. did not help.
Using Oauth 1 or 2 yields with error. I am fed up with that. Perhaps
in time there's going to be a better solution, something will be fixed
or someone will make a YT video showcasing how to do it correctly :shrug: